Consolidate popup code into PopupClient
