+++ lib/msal-browser/src/interaction_handler/InteractionHandler.ts
@@ -14,32 +14,26 @@ export type InteractionParams = {};
 /**
  * Abstract class which defines operations for a browser interaction handling class.
  */
-export abstract class InteractionHandler {
+export class InteractionHandler {
 
     protected authModule: AuthorizationCodeClient;
     protected browserStorage: BrowserCacheManager;
     protected authCodeRequest: CommonAuthorizationCodeRequest;
-    protected browserRequestLogger: Logger;
+    protected logger: Logger;
 
-    constructor(authCodeModule: AuthorizationCodeClient, storageImpl: BrowserCacheManager, authCodeRequest: CommonAuthorizationCodeRequest, browserRequestLogger: Logger) {
+    constructor(authCodeModule: AuthorizationCodeClient, storageImpl: BrowserCacheManager, authCodeRequest: CommonAuthorizationCodeRequest, logger: Logger) {
         this.authModule = authCodeModule;
         this.browserStorage = storageImpl;
         this.authCodeRequest = authCodeRequest;
-        this.browserRequestLogger = browserRequestLogger;
+        this.logger = logger;
     }
-
-    /**
-     * Function to enable user interaction.
-     * @param requestUrl
-     */
-    abstract initiateAuthRequest(requestUrl: string, params: InteractionParams): Window | Promise<HTMLIFrameElement> | Promise<void>;
 
     /**
      * Function to handle response parameters from hash.
      * @param locationHash
      */
     async handleCodeResponseFromHash(locationHash: string, state: string, authority: Authority, networkModule: INetworkModule): Promise<AuthenticationResult> {
-        this.browserRequestLogger.verbose("InteractionHandler.handleCodeResponse called");
+        this.logger.verbose("InteractionHandler.handleCodeResponse called");
         // Check that location hash isn't empty.
         if (StringUtils.isEmpty(locationHash)) {
             throw BrowserAuthError.createEmptyHashError(locationHash);
@@ -76,7 +70,7 @@ export abstract class InteractionHandler {
      * @returns 
      */
     async handleCodeResponseFromServer(authCodeResponse: AuthorizationCodePayload, state: string, authority: Authority, networkModule: INetworkModule, validateNonce: boolean = true): Promise<AuthenticationResult> {
-        this.browserRequestLogger.trace("InteractionHandler.handleCodeResponseFromServer called");
+        this.logger.trace("InteractionHandler.handleCodeResponseFromServer called");
 
         // Handle code response.
         const stateKey = this.browserStorage.generateStateKey(state);
