+++ lib/msal-browser/src/request/PopupWindowAttributes.ts
@@ -1,0 +1,22 @@
+/*
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+/**
+ * Popup configurations for setting dimensions and position of popup window
+ */
+export type PopupWindowAttributes = {
+    popupSize?: PopupSize,
+    popupPosition?: PopupPosition
+};
+
+export type PopupSize = {
+    height: number;
+    width: number;
+};
+
+export type PopupPosition = {
+    top: number;
+    left: number;
+};
